The main objective of this PR is to simplify handling of subcomponents (#421) and to make them immediately traverseable to serve model building and other tasks (#536). To this end, the following principles were instituted:

1. Components own their subcomponents.
 - a Component owns it properties and properties can be components: these are subcomponents.
   - user/developer *never* has to mark components that are properties as subcomponents
 - components that are data members of a Component are also its subcomponents (also do not need to be marked)
 - any other owned Component can be marked as a subcomponent (e.g. Model generated loop breaking constraints)

2. The ability to iterate through components is only dependent on the subcomponents list

**The main priorities for review are:**

1. Changes to Component so that:
 - components that are properties are listed as subcomponents automatically 
 - subcomponents (and all other structural lists) are copied appropriately
 - no requirement to connect() to traverse through subcomponents 

2. Update to Model:
 - removed the requirement to connect() in Model::extendFinalizeFromProperties()
 - subcomponent list and multibody tree are decoupled (tree building can now be delayed to connecting stage)
 - adding a Component assumes model takes ownership of it and it is a member of a property list.

3. Related test cases: testComponentInterface, testSerializeOpenSimObjects, testComponents, testIterators to verify subcomponents (as a property and internal member) are handled correctly and are traverseable. 

4. Additional Intention/consequence:
  -  always finalize Component properties upon construction (if Components are properties they need to be properly in place first). Only exception is for deserialization testing where we know there isn't a valid model there and we only care about comparing property values not their validity. 

5. Known Hacks/Bugs
 - I have not dealt the with the ComponentList to traverse subcomponents. To enable getting a valid ComponentList at anytime I force the creation of Component's traversal tree whenever a ComponentList is asked for. I put in some ugly const casting that MUST be removed in a subsequent PR, but that is independent of the the current changes. 
 - The Model leaks secondarily created Components needed to create a valid multibody tree. That is instead of adding to properties (and thus the serialized version of the Model) I just mark it as a subcomponent but don't hang on to it in order to delete it. There are several easy options to fix this but wanted to discuss separately and not delay this PR.

The remainder of the changes are to handle the fallout of the above changes, which were many. I also apologize for any spurious changes made as cleanup while debugging and testing.

Thanks and happy reviewing!