@aymanhab 2016-01-22
@aseth1 Thanks for the in person explanation yesterday regarding parent/ownership language. A few high level items based on reading the description, and Component.cpp:
1.  A good test for a hierarchical model building is to be able to take a model and another model and combine them into one model (maybe sharing ground only). Is that possible after this PR with real models?
2.  "Adding a component causes transfer of ownership", did anything change in the interface for how it's done. In the past we had to tell the scripting/SWIG about that so memory management doesn't blow up for scripting. This could be unrelated to this PR, but needs handling in SWIG. A related issue is handling stack allocated objects and adding them to the model.
3.  Does this PR handle or do the plumbing to allow for "internal" unexposed components. I'm interested because the "frame-geometry" can possibly use this functionality if available (a separate PR of course).
4.  Recomputing the tree-traversal on every call to getComponentList is expensive and a bit excessive especially that the tree doesn't change at all post initialization. A caching mechanism can be implemented to improve performance but I still think this can be optimized if you know exactly when components are added to the tree, so there's no cost post model building.
5.  Everything here works with Components while models have been made of ModelComponents, is there a long term future for this distinction?
6.  The ownership paradigm implies no sharing/re-reuse, what happens when a user tries to add a component already owned? Does the user need to "new" components to add them to the model? does the ownership switch again?

Thanks much, sorry for the long list, it maybe better to discuss in person if you feel github PR is not the right medium.