@aseth1 2016-01-22
@aymanhab Thank you for the nice list of issues. I am going to repost most of these as an issue for further discussion outside this PR. The objective for this PR was narrower in scope and did not specifically target many of these, but I can provide more details how each of the items were impacted.

>1.A good test for a hierarchical model building is to be able to take a model and another model and combine them into one model (maybe sharing ground only). Is that possible after this PR with real models?

I agree and it is definitely a capability I want to work in OpenSim. The Component interface supports this, see [here](https://github.com/opensim-org/opensim-core/pull/740/files#diff-074825ae2a365226376a3224512e6329R533). But,with current models it will not work. One issue there is that model owns and expects to own the system. Multiple Ground components should be OK since they all would reference the same multibody system Ground and it would allow each model to have its own Frames (e.g. offsets) and Geometry associated with Ground that the other or an Uber model would have to know. For the time being it is a relatively low priority but we are making steady progress in that direction.

>2."Adding a component causes transfer of ownership", did anything change in the interface for how it's done. In the past we had to tell the scripting/SWIG about that so memory management doesn't blow up for scripting. This could be unrelated to this PR, but needs handling in SWIG. A related issue is handling stack allocated objects and adding them to the model.

No change here whatsover. The same issues remain and warrant a separate discussion.

>3.Does this PR handle or do the plumbing to allow for "internal" unexposed components. I'm interested because the "frame-geometry" can possibly use this functionality if available (a separate PR of course).

Yes and no. If you have an new member that is a Component and use the new constructor to initialize it like [this](https://github.com/opensim-org/opensim-core/pull/740/files#diff-074825ae2a365226376a3224512e6329R118) then you are done. It will be a subcomponent and it will have Component calls propagated to it. If you need to create Components on the fly due what happens after finalizeFromProperties (e.g realize you need to break a Loop in the case of Model) then you still have to manually markAsSubcomponent(). My sense is that this will be so rarely used as to not warrant the effort. You could easily manage it for Frame's builtin FrameGeometry to keep it a private component of Frame.

>4.Recomputing the tree-traversal on every call to getComponentList is expensive and a bit excessive especially that the tree doesn't change at all post initialization. A caching mechanism can be implemented to improve performance but I still think this can be optimized if you know exactly when components are added to the tree, so there's no cost post model building.

Absolutely! That is why I listed it in Hack/Bugs. It is not intended to stay as is. It was shortest path to show that the interface works. We can correct and optimize the backend as necessary. I am not as concerned about the inefficiency as with the guarantee of correctness with other moving parts. For all its flaws it at least guarantees that the list is up-to-date and there is no order dependency, flags, or side-effects to worry about! 

>5.Everything here works with Components while models have been made of ModelComponents, is there a long term future for this distinction?

For the time being you can use them synonymously. The distinction will be more clear when there are Components that are not ModelComponents. That is they are purely calculating Components (still need system resources). These would be things like Filters, Signal generators, Multiplexers  and mathematical operators.

> 6.The ownership paradigm implies no sharing/re-reuse, what happens when a user tries to add a component already owned? Does the user need to "new" components to add them to the model? does the ownership switch again?

Same as 2. it behaves as it does now, unfortunately- failing during destruction when the second Set tries to delete its objects. This is a problem and it is not being addressed here. It could be remedied by using a data structure of smart pointers (instead of or inside of Sets) or use handles - but that is for another discussion/fix/PR.