@sherm1 2016-01-25
Discussed with @aseth1. We decided that the best approach would be: merge this big PR (after whatever local changes Ajay wants to make); make an issue with a checklist for remaining changes; then do a series of small, focused, easier-to-review PRs with remaining changes.