@chrisdembia 2016-01-25
It seems preferable to reverse that process: first make small PRs that are necessary for an upcoming feature, then make a PR for the feature. Is that possible here?