@aseth1 2016-01-25
@chrisdembia that is the ideal but it isn't always possible. For example, a small change in the call sequence of finalizeFromProperties caused many fixes. I feel this is a relatively small/simple PR with lots of changes and testing to accommodate it which accounts for the bulk of the files involved. Now that the bulk change is done I can focus more on the correct implementation of the iterator and it won't get buried.