@aseth1 2016-01-25
Appveyor
Recent:
build (2141)
44/68 Test #47: testIterators ................................   Passed    2.14 sec
 build (2166)
43/68 Test #47: testIterators ................................   Passed    1.20 sec

This branch (2125):
45/68 Test #47: testIterators ................................   Passed    1.59 sec

Gcc Travis:
Previous (2837.2)
38/66 Test #47: testIterators ................................   Passed    0.32 sec
Recent (2873.2)
37/66 Test #47: testIterators ................................   Passed    1.82 sec 

This branch (2840.2)
37/66 Test #47: testIterators ................................   Passed    0.51 sec

My machine:
(master): 48/66 Test #47: testIterators ................................   Passed    0.37 sec
(branch): 5/66 Test #47: testIterators ................................   Passed    0.14 sec
(No these are not reversed.)

There seems to be little indication that change slows down performance but there is a lot of variability.